Refactor span line calculations into dedicated functions; add unimplemented tests for remaining nom input traits for span

+++ vimwiki/src/lang/parsers/errors.rs
@@ -8,7 +8,7 @@ pub struct LangParserError {
     ctx: String,
     sample: String,
     offset: usize,
-    line: u32,
+    line: usize,
     column: usize,
     next: Option<Box<Self>>,
 }
